+++ models/src/notification_settings.rs
@@ -4,7 +4,7 @@ use sea_orm::entity::prelude::*;
 use sea_orm::{Set, ActiveModelTrait};
 use uuid::Uuid;
 
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Serialize, Deserialize)]
 #[serde(crate = "rocket::serde")]
 #[sea_orm(table_name = "notification_settings")]
 pub struct Model {
@@ -14,10 +14,10 @@ pub struct Model {
     pub time_based_enabled: bool,
     pub time_threshold_minutes: Option<i32>,
     pub earnings_based_enabled: bool,
-    pub earnings_threshold: Option<Decimal>,
+    pub earnings_threshold: Option<f64>,
     pub currency: Option<String>,
     pub daily_goal_enabled: bool,
-    pub daily_hours_goal: Option<Decimal>,
+    pub daily_hours_goal: Option<f64>,
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
 }
